requirements.txt
Flask==0.12.2
gunicorn==19.7.1
python-dotenv==0.7.1
nltk==3.2.5
coverage==4.4.1